+++ package.json
@@ -87,8 +87,8 @@
     "npm-run-all": "^4.1.2",
     "prettier": "^1.19.1",
     "prismjs": "^1.5.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-error-boundary": "^1.2.5",
     "react-router-dom": "^5.1.2",
     "react-values": "^0.3.0",
