+++ packages/slate-react/src/components/restore-dom/restore-dom-manager.ts
@@ -1,6 +1,6 @@
 import { RefObject } from 'react'
 import { ReactEditor } from '../../plugin/react-editor'
-import { isTrackedMutation } from './use-mutation-observer'
+import { isTrackedMutation } from '../../utils/dom'
 
 export type RestoreDOMManager = {
   registerMutations: (mutations: MutationRecord[]) => void
