+++ site/pages/examples/[example].tsx
@@ -265,7 +265,6 @@ const ExamplePage = ({ example }: { example: string }) => {
       onError={(error, stacktrace) => {
         setError(error)
         setStacktrace(stacktrace)
-        console.error(error)
       }}
     >
       <div>
